Create C3 system watchdog handling service
This is a Rust implementation of oresat-c3-software/oresat_watchdog and
pet-wdt. It manages the hardware watchdog, listening on UDP port 20001
for a heartbeat and if it doesn't hear anything after a specified
timeout, will quit and let the hardware watchdog tick over.

It also is capable of generating a debian package to install itself,
intended for installing in the C3 flight image.

I wanted to keep the implementation fairly close to how it would look in
C/Linux, specifically avoiding the async ecosystem for now. This uses
nix for signalfd/timerfd, gpiod for gpio, and mio for poll/the event
loop. I used timerfd in ax5043, but that's built on top of rustix, a
competitor to nix, and I needed nix directly for signalfd so I also used
its timerfd to keep dependencies small. nix's timerfd implementation is
closer to the C interface so it's not as nice as the timerfd crate.