Pull socket/timer out into Pingee, misc refactor
The Pingee type I feel better emphasizes the connection between the
socket and the timeout timer.

This also removes impl event::Source for Petter. As fun as it was to
have that impl, I only ever used it to save a few characters on one line
so the whole thing didn't really seem worth it.