Use cfg card=c3 to run on flatsat
As a quick hack runner.sh always ran the arm builds on the C3, but this
doesn't play super nicely with other systems. If for example I wanted to
run in a VM or on a rasberry pi, I'd have to manually configure it.
Using cfg checks makes this a much more flexible process.